+++ python/ambassador/envoy/v2/v2httpfilter.py
@@ -89,7 +89,7 @@ def V2HTTPFilter(irfilter: IRFilter, v2config: 'V2Config'):
     if irfilter.kind == 'IRAuth':
         if irfilter.api_version == 'getambassador.io/v0':
             return 'IRAuth_v0'
-        elif (irfilter.api_version == 'getambassador.io/v1') or (irfilter.api_version == 'getambassador.io/v2'):
+        elif irfilter.api_version in ['getambassador.io/v1', 'getambassador.io/v2']:
             return 'IRAuth_v1-2'
         else:
             irfilter.post_error('AuthService version %s unknown, treating as v2' % irfilter.api_version)
@@ -238,7 +238,7 @@ def V2HTTPFilter_authv1(auth: IRAuth, v2config: 'V2Config'):
     assert auth.cluster
     cluster = typecast(IRCluster, auth.cluster)
 
-    if (auth.api_version != "getambassador.io/v1") and (auth.api_version != "getambassador.io/v2"):
+    if auth.api_version not in ['getambassador.io/v1', 'getambassador.io/v2']:
         auth.ir.logger.warning("IRAuth_v1 working on %s, mismatched at %s" % (auth.name, auth.api_version))
 
     assert auth.proto
