+++ python/ambassador/envoy/v3/v3bootstrap.py
@@ -1,5 +1,6 @@
 from typing import TYPE_CHECKING
 from typing import cast as typecast
+
 import os
 
 from ...ir.ircluster import IRCluster
@@ -113,6 +114,13 @@ class V3Bootstrap(dict):
             if config.ir.statsd['dogstatsd']:
                 name = 'envoy.stat_sinks.dog_statsd'
                 typename = 'type.googleapis.com/envoy.config.metrics.v3.DogStatsdSink'
+                dd_entity_id = os.environ.get('DD_ENTITY_ID', None)
+                if dd_entity_id:
+                    stats_tags = self.setdefault('stats_config', {}).setdefault('stats_tags', [])
+                    stats_tags.append({
+                        'tag_name': 'dd.internal.entity_id',
+                        'fixed_value': dd_entity_id
+                    })
             else:
                 name = 'envoy.stats_sinks.statsd'
                 typename = 'type.googleapis.com/envoy.config.metrics.v3.StatsdSink'
